+++ controllers/homeRoutes.js
@@ -1,24 +1,42 @@
 const router = require('express').Router();
-const { User } = require('../models');
+const { User, Posts } = require('../models');
 const withAuth = require('../utils/auth');
 
 router.get('/', withAuth, async (req, res) => {
   try {
-    const userData = await User.findAll({
-      attributes: { exclude: ['password'] },
-      order: [['name', 'ASC']],
+    const userPost = await Posts.findAll({
+        include: [{ model: User, attributes: { exclude: ['password'] } }]
     });
 
-    const users = userData.map((project) => project.get({ plain: true }));
-
+    const posts = userPost.map((project) => project.get({ plain: true }));
+    console.log(posts);
+    
     res.render('homepage', {
-      users,
-      logged_in: req.session.logged_in,
-    });
+            posts,
+            logged_in: req.session.logged_in,
+          });
   } catch (err) {
-    res.status(500).json(err);
+      res.status(500).json(err)
   }
 });
+
+// router.get('/', async (req, res) => {
+//   try {
+//     const userData = await User.findAll({
+//       attributes: { exclude: ['password'] },
+//       order: [['name', 'ASC']],
+//     });
+
+//     const users = userData.map((project) => project.get({ plain: true }));
+
+//     res.render('homepage', {
+//       users,
+//       logged_in: req.session.logged_in,
+//     });
+//   } catch (err) {
+//     res.status(500).json(err);
+//   }
+// });
 
 router.get('/login', (req, res) => {
   if (req.session.logged_in) {
