+++ views/homepage.handlebars
@@ -1,4 +1,4 @@
-<!DOCTYPE html>
+{{!-- <!DOCTYPE html>
 <html>
 
 <head>
@@ -6,19 +6,27 @@
   <style>
     /* Add your CSS styles here */
   </style>
-</head>
+</head> --}}
 
-<body>
+{{!-- <body> --}}
   <h1>Welcome to the Homepage</h1>
 
-  <div id="feed">
-    {{#each posts}}
-    <div class="post">
-      <h2>{{title}}</h2>
-      <p>{{content}}</p>
+  {{!-- <div id="feed"> --}}
+
+{{!-- Card should be a partial --}}
+    {{#each posts as |post|}}
+    <div class="row align-center mb-5">
+      <div class="col-md-6">
+        <h2>{{user.name}}</h2>
+      <p>{{post.body}}</p>
+      <p>{{user.linkedin}}</p>
+      <p>{{user.github}}</p>
+      <p>{{user.email}}</p>
+      </div>
     </div>
     {{/each}}
-  </div>
+
+  {{!-- </div> --}}
 
   <button id="showMoreButton">Show 5 More</button>
 
@@ -37,6 +45,6 @@
   </div>
 
   <script src="homepage.js"></script>
-</body>
+{{!-- </body>
 
-</html>
+</html> --}}
